+++ src/main/java/io/xpring/ilp/IlpException.java
@@ -1,0 +1,25 @@
+package io.xpring.ilp;
+
+public class IlpException extends Exception {
+    /** The type of exception. */
+    private IlpExceptionType type;
+
+    /**
+     * Create a new exception.
+     *
+     * @param type The type of exception.
+     * @param message The message to to include in the exception
+     */
+    public IlpException(IlpExceptionType type, String message) {
+        super(message);
+
+        this.type = type;
+    }
+
+    /**
+     * @return The exception type.
+     */
+    public IlpExceptionType getType() {
+        return this.type;
+    }
+}