+++ src/main/java/io/xpring/xrpl/XRPClientDecorator.java
@@ -11,9 +11,9 @@ public interface XRPClientDecorator {
      *
      * @param xrplAccountAddress The X-Address to retrieve the balance for.
      * @return A {@link BigInteger} with the number of drops in this account.
-     * @throws XpringException If the given inputs were invalid.
+     * @throws XRPException If the given inputs were invalid.
      */
-    BigInteger getBalance(final String xrplAccountAddress) throws XpringException;
+    BigInteger getBalance(final String xrplAccountAddress) throws XRPException;
 
 
     /**
@@ -25,7 +25,7 @@ public interface XRPClientDecorator {
      * @param transactionHash The hash of the transaction.
      * @return The status of the given transaction.
      */
-    public TransactionStatus getPaymentStatus(String transactionHash) throws XpringException;
+    public TransactionStatus getPaymentStatus(String transactionHash) throws XRPException;
 
     /**
      * Transact XRP between two accounts on the ledger.
@@ -34,20 +34,20 @@ public interface XRPClientDecorator {
      * @param destinationAddress The X-Address to send the XRP to.
      * @param sourceWallet The {@link Wallet} which holds the XRP.
      * @return A transaction hash for the payment.
-     * @throws XpringException If the given inputs were invalid.
+     * @throws XRPException If the given inputs were invalid.
      */
     String send(
             final BigInteger amount,
             final String destinationAddress,
             final Wallet sourceWallet
-    ) throws XpringException;
+    ) throws XRPException;
 
     /**
      * Retrieve the latest validated ledger sequence on the XRP Ledger.
      *
      * @return A long representing the sequence of the most recently validated ledger.
      */
-    int getLatestValidatedLedgerSequence() throws XpringException;
+    int getLatestValidatedLedgerSequence() throws XRPException;
 
     /**
      * Retrieve the raw transaction status for the given transaction hash.
@@ -55,7 +55,7 @@ public interface XRPClientDecorator {
      * @param transactionHash: The hash of the transaction.
      * @return an {@link RawTransactionStatus} containing the raw transaction status.
      */
-    RawTransactionStatus getRawTransactionStatus(String transactionHash) throws XpringException;
+    RawTransactionStatus getRawTransactionStatus(String transactionHash) throws XRPException;
 
     /**
      * Check if an address exists on the XRP Ledger.
@@ -63,5 +63,5 @@ public interface XRPClientDecorator {
      * @param address The address to check the existence of.
      * @return A boolean if the account is on the XRPLedger.
      */
-    boolean accountExists(String address) throws XpringException;
+    boolean accountExists(String address) throws XRPException;
 }