+++ src/main/java/io/xpring/xrpl/javascript/JavaScriptWalletFactory.java
@@ -1,7 +1,8 @@
 package io.xpring.xrpl.javascript;
 
 import io.xpring.xrpl.Utils;
-import io.xpring.xrpl.XpringException;
+import io.xpring.xrpl.XRPException;
+import io.xpring.xrpl.XRPExceptionType;
 import org.graalvm.polyglot.Context;
 import org.graalvm.polyglot.PolyglotException;
 import org.graalvm.polyglot.Value;
@@ -62,13 +63,13 @@ public class JavaScriptWalletFactory {
      * @param publicKey A hex encoded string representing the public key.
      * @param privateKey A hex encoded string representing the private key.
      * @param isTest Whether the address is for use on a test network.
-     * @throws XpringException If either input key is malformed.
+     * @throws XRPException If either input key is malformed.
      * @return A new {@link JavaScriptWallet}.
      */
-    public JavaScriptWallet walletFromKeys(String publicKey, String privateKey, boolean isTest) throws XpringException {
+    public JavaScriptWallet walletFromKeys(String publicKey, String privateKey, boolean isTest) throws XRPException {
         Value wallet = this.wallet.newInstance(publicKey, privateKey, isTest);
         if (wallet.isNull()) {
-            throw new XpringException("Invalid inputs");
+            throw new XRPException(XRPExceptionType.INVALID_INPUTS, "Invalid inputs");
         }
         return new JavaScriptWallet(wallet);
     }
@@ -78,13 +79,13 @@ public class JavaScriptWalletFactory {
      *
      * @param seed A base58check encoded seed for the wallet.
      * @param isTest Whether the address is for use on a test network.
-     * @throws XpringException If the seed is malformed.
+     * @throws XRPException If the seed is malformed.
      * @return A new {@link JavaScriptWallet}.
      */
-    public JavaScriptWallet walletFromSeed(String seed, boolean isTest) throws XpringException {
+    public JavaScriptWallet walletFromSeed(String seed, boolean isTest) throws XRPException {
         Value wallet = this.wallet.invokeMember("generateWalletFromSeed", seed, isTest);
         if (wallet.isNull()) {
-            throw new XpringException("Invalid Seed");
+            throw new XRPException(XRPExceptionType.INVALID_INPUTS, "Invalid Seed");
         }
         return new JavaScriptWallet(wallet);
     }
@@ -95,23 +96,23 @@ public class JavaScriptWalletFactory {
      * @param mnemonic       A space separated mnemonic.
      * @param derivationPath A derivation. If null, the default derivation path will be used.
      * @param isTest Whether the address is for use on a test network.
-     * @throws XpringException If the mnemonic or derivation path are malformed.
+     * @throws XRPException If the mnemonic or derivation path are malformed.
      * @return A new {@link JavaScriptWallet}.
      */
-    public JavaScriptWallet walletFromMnemonicAndDerivationPath(String mnemonic, String derivationPath, boolean isTest) throws XpringException {
+    public JavaScriptWallet walletFromMnemonicAndDerivationPath(String mnemonic, String derivationPath, boolean isTest) throws XRPException {
         try {
             String normalizedDerivationPath = derivationPath != null ? derivationPath : this.getDefaultDerivationPath();
             Value wallet = this.wallet.invokeMember("generateWalletFromMnemonic", mnemonic, normalizedDerivationPath, isTest);
 
             if (wallet.isNull()) {
-                throw new XpringException(invalidMnemonicOrDerivationPathMessage);
+                throw new XRPException(XRPExceptionType.INVALID_INPUTS, invalidMnemonicOrDerivationPathMessage);
             }
 
             return new JavaScriptWallet(wallet);
         } catch (PolyglotException exception) {
-            throw new XpringException(invalidMnemonicOrDerivationPathMessage);
+            throw new XRPException(XRPExceptionType.INVALID_INPUTS, invalidMnemonicOrDerivationPathMessage);
         } catch (JavaScriptLoaderException exception) {
-            throw new XpringException(invalidMnemonicOrDerivationPathMessage);
+            throw new XRPException(XRPExceptionType.INVALID_INPUTS, invalidMnemonicOrDerivationPathMessage);
         }
     }