+++ src/test/java/io/xpring/xrpl/DefaultXRPClientTest.java
@@ -60,7 +60,7 @@ public class DefaultXRPClientTest {
     private static final BigInteger AMOUNT = new BigInteger("1");
 
     @Test
-    public void getBalanceTest() throws IOException, XpringException {
+    public void getBalanceTest() throws IOException, XRPException {
         // GIVEN a DefaultXRPClient with mocked networking which will succeed.
         DefaultXRPClient client = getClient();
 
@@ -72,18 +72,18 @@ public class DefaultXRPClientTest {
     }
 
     @Test
-    public void getBalanceWithClassicAddressTest() throws IOException, XpringException {
+    public void getBalanceWithClassicAddressTest() throws IOException, XRPException {
         // GIVEN a classic address.
         ClassicAddress classicAddress = Utils.decodeXAddress(XRPL_ADDRESS);
         DefaultXRPClient client = getClient();
 
         // WHEN the balance for the classic address is retrieved THEN an error is thrown.
-        expectedException.expect(XpringException.class);
+        expectedException.expect(XRPException.class);
         client.getBalance(classicAddress.address());
     }
 
     @Test
-    public void getBalanceTestWithFailedAccountInfo() throws IOException, XpringException {
+    public void getBalanceTestWithFailedAccountInfo() throws IOException, XRPException {
         // GIVEN a XRPClient with mocked networking which will fail to retrieve account info.
         Result<GetAccountInfoResponse, Throwable> accountInfoResult = Result.error(GENERIC_ERROR);
         DefaultXRPClient client = getClient(
@@ -100,7 +100,7 @@ public class DefaultXRPClientTest {
 
 
     @Test
-    public void paymentStatusWithUnvalidatedTransactionAndFailureCode() throws IOException, XpringException {
+    public void paymentStatusWithUnvalidatedTransactionAndFailureCode() throws IOException, XRPException {
         // Iterate over different types of transaction status codes which represent failures.
         for (String transactionFailureCode : TRANSACTION_FAILURE_STATUS_CODES) {
             // GIVEN a XRPClient which will return an unvalidated transaction with a failed code.
@@ -120,7 +120,7 @@ public class DefaultXRPClientTest {
     }
 
     @Test
-    public void paymentStatusWithUnvalidatedTransactionAndSuccessCode() throws IOException, XpringException {
+    public void paymentStatusWithUnvalidatedTransactionAndSuccessCode() throws IOException, XRPException {
         // GIVEN a XRPClient which will return an unvalidated transaction with a success code.
         DefaultXRPClient client = getClient(
                 Result.ok(makeGetAccountInfoResponse(DROPS_OF_XRP_IN_ACCOUNT)),
@@ -137,7 +137,7 @@ public class DefaultXRPClientTest {
     }
 
     @Test
-    public void paymentStatusWithValidatedTransactionAndFailureCode() throws IOException, XpringException {
+    public void paymentStatusWithValidatedTransactionAndFailureCode() throws IOException, XRPException {
         // Iterate over different types of transaction status codes which represent failures.
         for (String transactionFailureCode : TRANSACTION_FAILURE_STATUS_CODES) {
             // GIVEN a XRPClient which will return an validated transaction with a failed code.
@@ -157,7 +157,7 @@ public class DefaultXRPClientTest {
     }
 
     @Test
-    public void paymentStatusWithValidatedTransactionAndSuccessCode() throws IOException, XpringException {
+    public void paymentStatusWithValidatedTransactionAndSuccessCode() throws IOException, XRPException {
         // GIVEN a XRPClient which will return an validated transaction with a success code.
         DefaultXRPClient client = getClient(
                 Result.ok(makeGetAccountInfoResponse(DROPS_OF_XRP_IN_ACCOUNT)),
@@ -174,7 +174,7 @@ public class DefaultXRPClientTest {
     }
 
     @Test
-    public void paymentStatusWithNodeError() throws IOException, XpringException {
+    public void paymentStatusWithNodeError() throws IOException, XRPException {
         // GIVEN a XRPClient which will error when a transaction status is requested..
         DefaultXRPClient client = getClient(
                 Result.ok(makeGetAccountInfoResponse(DROPS_OF_XRP_IN_ACCOUNT)),
@@ -189,7 +189,7 @@ public class DefaultXRPClientTest {
     }
 
     @Test
-    public void submitTransactionTest() throws IOException, XpringException {
+    public void submitTransactionTest() throws IOException, XRPException {
         // GIVEN a XRPClient with mocked networking which will succeed.
         DefaultXRPClient client = getClient();
         Wallet wallet = new Wallet(WALLET_SEED);
@@ -202,19 +202,19 @@ public class DefaultXRPClientTest {
     }
 
     @Test
-    public void submitTransactionWithClassicAddress() throws IOException, XpringException {
+    public void submitTransactionWithClassicAddress() throws IOException, XRPException {
         // GIVEN a classic address.
         DefaultXRPClient client = getClient();
         ClassicAddress classicAddress = Utils.decodeXAddress(XRPL_ADDRESS);
         Wallet wallet = new Wallet(WALLET_SEED);
 
         // WHEN XRP is sent to the classic address THEN an error is thrown.
-        expectedException.expect(XpringException.class);
+        expectedException.expect(XRPException.class);
         client.send(AMOUNT, classicAddress.address(), wallet);
     }
 
     @Test
-    public void submitTransactionWithFailedAccountInfo() throws IOException, XpringException {
+    public void submitTransactionWithFailedAccountInfo() throws IOException, XRPException {
         // GIVEN a XRPClient which will fail to return account info.
         Result<GetAccountInfoResponse, Throwable> accountInfoResult = Result.error(GENERIC_ERROR);
         DefaultXRPClient client = getClient(
@@ -231,7 +231,7 @@ public class DefaultXRPClientTest {
     }
 
     @Test
-    public void submitTransactionWithFailedFee() throws IOException, XpringException {
+    public void submitTransactionWithFailedFee() throws IOException, XRPException {
         // GIVEN a XRPClient which will fail to retrieve a fee.
         Result<GetFeeResponse, Throwable> feeResult = Result.error(GENERIC_ERROR);
         DefaultXRPClient client = getClient(
@@ -249,7 +249,7 @@ public class DefaultXRPClientTest {
 
 
     @Test
-    public void submitTransactionWithFailedSubmit() throws IOException, XpringException {
+    public void submitTransactionWithFailedSubmit() throws IOException, XRPException {
         // GIVEN a XRPClient which will fail to submit a transaction.
         Result<SubmitTransactionResponse, Throwable> submitResult = Result.error(GENERIC_ERROR);
         DefaultXRPClient client = getClient(
@@ -266,7 +266,7 @@ public class DefaultXRPClientTest {
     }
 
     @Test
-    public void accountExistsTest() throws IOException, XpringException {
+    public void accountExistsTest() throws IOException, XRPException {
         // GIVEN a DefaultXRPClient with mocked networking which will succeed.
         DefaultXRPClient client = getClient();
 
@@ -278,18 +278,18 @@ public class DefaultXRPClientTest {
     }
 
     @Test
-    public void accountExistsWithClassicAddressTest() throws IOException, XpringException {
+    public void accountExistsWithClassicAddressTest() throws IOException, XRPException {
         // GIVEN a classic address.
         ClassicAddress classicAddress = Utils.decodeXAddress(XRPL_ADDRESS);
         DefaultXRPClient client = getClient();
 
         // WHEN the existence of the account is checked for the classic address THEN an error is thrown.
-        expectedException.expect(XpringException.class);
+        expectedException.expect(XRPException.class);
         client.accountExists(classicAddress.address());
     }
 
     @Test
-    public void accountExistsTestWithNotFoundError() throws IOException, XpringException {
+    public void accountExistsTestWithNotFoundError() throws IOException, XRPException {
         // GIVEN a XRPClient with mocked networking which will fail to retrieve account info w/ NOT_FOUND error code.
         StatusRuntimeException notFoundError = new StatusRuntimeException(Status.NOT_FOUND);
         Result<GetAccountInfoResponse, Throwable> accountInfoResult = Result.error(notFoundError);
@@ -308,7 +308,7 @@ public class DefaultXRPClientTest {
     }
 
     @Test
-    public void accountExistsTestWithUnkonwnError() throws IOException, XpringException {
+    public void accountExistsTestWithUnkonwnError() throws IOException, XRPException {
         // GIVEN a XpringClient with mocked networking which will fail to retrieve account info w/ UNKNOWN error code.
         StatusRuntimeException notFoundError = new StatusRuntimeException(Status.UNKNOWN);
         Result<GetAccountInfoResponse, Throwable> accountInfoResult = Result.error(notFoundError);